requirements.txt
# fastapi[all]

fastapi==0.136.0

pydantic==2.9.0
# pydantic==2.8.0
pydantic-settings==2.4.0
python-dotenv==1.0.0
# pysqlite3-binary

beanie==1.26.0

langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-core==0.3.0
langchain-experimental==0.3.0
langchain-openai==0.2.0
langchain-chroma==0.1.4
langchain-cli==0.0.32
# twilio
python-multipart==0.0.22
# pinecone-client==6.0.0
pypdf
# onnxruntime==1.17.0
python-telegram-bot==22.7

